[TD-16-km] main 함수를 분리하자
